pyproject.toml
[project]
name = "mcp-stock"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.9.4",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
